+++ language_bindings/python/bind_py_module.cc
@@ -46,6 +46,7 @@ PYBIND11_MODULE(_muGrid, mod) {
   add_state_field_classes(mod);
   add_field_collection_classes(mod);
   add_options_dictionary(mod);
+  add_testing(mod);
 #ifdef WITH_NETCDF_IO
   add_file_io_classes(mod);
 #endif